// web/src/lib/api/generated/models/peopleCodexProfileRequest.ts
/**
 * Generated by Orval. Do not edit manually.
 */

export interface PeopleCodexProfileRequest {
  allow_sensitive: boolean;
  /** @minItems 1 */
  allowed_sources: string[];
  /** @minLength 1 */
  model: string;
  /** @minLength 1 */
  reasoning_effort: string;
  /** @minLength 1 */
  retention_posture: string;
  source_since: string;
  source_until?: string;
  /** @minLength 1 */
  training_posture: string;
}